refactor: restructure openai module with repository pattern
