+++ src/shared/queries/openai.queries.ts
@@ -1,0 +1,8 @@
+import { useMutation } from "@tanstack/react-query";
+import { openaiService } from "@shared/services/openai.service";
+
+export const useGeneratePrompt = () => {
+  return useMutation({
+    mutationFn: openaiService.generatePrompt,
+  });
+};
