+++ src/shared/repositories/openai.repository.ts
@@ -1,43 +1,68 @@
-import { openAIService } from "@shared/services/openai.service";
+import { httpClient } from "@shared/lib/http-client";
 
-export const openAIRepository = {
+export interface OpenAIPromptResponse {
+  suggestions: string[];
+}
+
+export const openaiRepository = {
+  generatePrompt: async (keywords: string): Promise<OpenAIPromptResponse> => {
+    const response = await httpClient("/openai/chat", {
+      method: "POST",
+      body: JSON.stringify({ keywords }),
+    });
+    return response as OpenAIPromptResponse;
+  },
+
   async chatCompletion(messages: any[]) {
-    try {
-      const response = await openAIService.chatCompletion(messages);
-      return response;
-    } catch (error) {
-      console.error("Error in repository chat completion:", error);
-      throw error;
-    }
+    const response = await httpClient("/openai/chat", {
+      method: "POST",
+      body: JSON.stringify({ messages }),
+    });
+    return response;
   },
 
   async generateResumeFeedback(resume: string) {
-    try {
-      const response = await openAIService.generateResumeFeedback(resume);
-      return response;
-    } catch (error) {
-      console.error("Error in repository generate resume feedback:", error);
-      throw error;
-    }
+    const messages = [
+      {
+        role: "system",
+        content: "You are a professional resume reviewer. Analyze the resume and provide constructive feedback.",
+      },
+      {
+        role: "user",
+        content: resume,
+      },
+    ];
+
+    return await this.chatCompletion(messages);
   },
 
   async generateInterviewQuestions(jobDescription: string, candidateProfile: string) {
-    try {
-      const response = await openAIService.generateInterviewQuestions(jobDescription, candidateProfile);
-      return response;
-    } catch (error) {
-      console.error("Error in repository generate interview questions:", error);
-      throw error;
-    }
+    const messages = [
+      {
+        role: "system",
+        content: "You are an experienced interviewer. Generate relevant interview questions based on the job description and candidate profile.",
+      },
+      {
+        role: "user",
+        content: `Job Description: ${jobDescription}\n\nCandidate Profile: ${candidateProfile}`,
+      },
+    ];
+
+    return await this.chatCompletion(messages);
   },
 
   async evaluateInterviewAnswer(question: string, answer: string) {
-    try {
-      const response = await openAIService.evaluateInterviewAnswer(question, answer);
-      return response;
-    } catch (error) {
-      console.error("Error in repository evaluate interview answer:", error);
-      throw error;
-    }
+    const messages = [
+      {
+        role: "system",
+        content: "You are an experienced interviewer. Evaluate the candidate's answer and provide constructive feedback.",
+      },
+      {
+        role: "user",
+        content: `Question: ${question}\n\nAnswer: ${answer}`,
+      },
+    ];
+
+    return await this.chatCompletion(messages);
   },
 };
